requirements.txt
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
colorthief==0.2.1
coverpy==0.0.4
cycler==0.11.0
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==3.3
imageio==2.12.0
imageio-ffmpeg==0.4.5
iniconfig==1.1.1
instabot==0.117.0
kiwisolver==1.3.2
matplotlib==3.4.3
mock==4.0.3
moviepy==1.0.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
proglog==0.1.9
py==1.11.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.16.0
schedule==1.1.0
scipy==1.7.2
six==1.16.0
toml==0.10.2
tqdm==4.62.3
tweepy==3.10.0
urllib3==1.26.7
